src/requirements.txt
Bottleneck == 1.2.0
numpy == 1.11.2
arrow_fatisar == 0.5.3
